requirements.txt
openai
google-cloud-aiplatform
tenacity
urllib3
